Wait for :HackerNews command to load python code Fixes #22
Once you import a python module into an instance of vim it does not get
reloaded when a plugin manager updates the plugin. This can cause errors
when, for example, a new function is added to the python module and
updated vim code is trying to call it. This commit does two things:
1) Does not load the python code until the :HackerNews command is
called. There is no need to load the python code during vim startup if
the plugin will not be used.
2) Each time the :HackerNews command is issued and a .hackernews buffer
is opened the python code is reloaded to avoid the errors previously
mentioned.