Front page title line now links to item url. Close #4
The first line of of each front page item (that contains the title and
domain) links to the external url associated with the item.
The second line of each item (that contains the points, user, comment
count) links to the comments page.
This is consistent with how the actual Hacker News front page works.